pyproject.toml
[tool.poetry]
name = "pokemon-card-app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
psycopg2-binary = "^2.9"
Pillow = "^10.0"
django-import-export = "^3.3"
django-filter = "^23.5"
django-tailwind = "^4.4.1"
django-htmx = "^1.27.0"
django-widget-tweaks = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"